requirements.txt
llama-cpp-python==0.2.63
streamlit==1.31.1
